Unforked S3 lib
