--- server/fork/deno-s3-lite-client/deps.ts
@@ -1,2 +1,0 @@
-export { readableStreamFromIterable } from "https://deno.land/std@0.170.0/streams/readable_stream_from_iterable.ts";
-export { Buffer } from "https://deno.land/std@0.170.0/io/buffer.ts";
