--- server/fork/deno-s3-lite-client/mod.ts
@@ -1,2 +1,0 @@
-export { Client as S3Client } from "./client.ts";
-export * as S3Errors from "./errors.ts";
